refactor: move CompilerProfiler into compilertop package as Profiler
Continues the breakup of the compiler-top TUI. Moves
util/CompilerProfiler.scala to tools/compilertop/Profiler.scala via
git mv (history preserved) and renames the class + companion to
Profiler — the `Compiler` prefix is redundant now that the file lives
in the compilertop package. Three consumer files updated:

  - api/Flix.scala       — explicit `import Profiler` added (couldn't
                           rely on the `util.*` wildcard anymore); 4
                           type references renamed.
  - util/CompilerTop.scala — DefnStats import re-pointed; constructor
                             param type renamed.
  - tools/compilertop/Aggregation.scala — DefnStats import re-pointed.

Pure code move + rename, no behavior change.
