+++ main/src/ca/uwaterloo/flix/api/Flix.scala
@@ -27,6 +27,7 @@ import ca.uwaterloo.flix.language.phase.optimizer.{LambdaDrop, Optimizer}
 import ca.uwaterloo.flix.language.{CompilationMessage, GenSym}
 import ca.uwaterloo.flix.runtime.CompilationResult
 import ca.uwaterloo.flix.tools.Summary
+import ca.uwaterloo.flix.tools.compilertop.Profiler
 import ca.uwaterloo.flix.util.*
 import ca.uwaterloo.flix.util.Formatter.NoFormatter
 import ca.uwaterloo.flix.util.collection.{Chain, MultiMap}
@@ -115,16 +116,16 @@ class Flix {
     * by [[setOptions]] when the compiler profiler is enabled; absent on the
     * default path so [[profile]] is a no-op with no measurement overhead.
     */
-  private var profiler: Option[CompilerProfiler] = None
+  private var profiler: Option[Profiler] = None
 
   /** The live compiler profiler TUI, if it has been started. */
   private var compilerTop: Option[CompilerTop] = None
 
   /** Returns the currently installed profiler, or `None`. */
-  def getProfiler: Option[CompilerProfiler] = profiler
+  def getProfiler: Option[Profiler] = profiler
 
   /** Installs (or removes, when `None`) the profiler that backs [[profile]]. */
-  def setProfiler(p: Option[CompilerProfiler]): Unit = profiler = p
+  def setProfiler(p: Option[Profiler]): Unit = profiler = p
 
   /**
     * Records the time spent running `thunk` against `sym`, attributed to
@@ -437,7 +438,7 @@ class Flix {
       throw new IllegalArgumentException("'opts' must be non-null.")
     options = opts
     if (opts.compilerTop && compilerTop.isEmpty) {
-      val p = new CompilerProfiler(() => getCurrentPhaseName)
+      val p = new Profiler(() => getCurrentPhaseName)
       setProfiler(Some(p))
       // The profiler subscribes to compiler events (e.g. NewConstraintsDef)
       // to track signals that are awkward to thread through `track()`.
